refactor(commentTest): commentRepositoryTest 클래스로 통합 및 데이터 정합성을 위해서 사용후 엔티티 삭제까지 처리
